Refactor archival
