+++ crates/parcel_core/src/types/utils.rs
@@ -1,3 +1,6 @@
+/// Implements bitflags archival using rkyv
+pub(crate) mod bitflags_archiver;
+
 #[macro_export]
 macro_rules! impl_bitflags_serde {
   ($t: ty) => {
